Remove namespace Dwelling + introduce index.ts
