+++ domain/furnishing/index.ts
@@ -1,0 +1,1 @@
+export { Dwelling } from "./Dwelling";
